+++ ha-business/src/main/java/jp/co/ha/business/db/crud/read/ApiCommunicationDataSearchService.java
@@ -23,5 +23,5 @@ public interface ApiCommunicationDataSearchService {
      *
      * @return トランザクションIDの最大値
      */
-    Integer selectLastTransactionId();
+    Long selectLastTransactionId();
 }
